Updates parser to handle HTML and units
Improves the Marketo template parser to correctly handle HTML content within tags and also include units of measure.

Escapes HTML entities in tag defaults unless 'allowhtml' is set to true. Adds support for displaying units specified in the 'unit' attribute.
Also modifies the extension to run automatically when the page loads instead of on a button click, and adds console logging for debugging.

Updates the version number to 0.3.